+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/ops-chooser.js
@@ -154,6 +154,16 @@ OpsExchangeDocument = Backbone.Model.extend({
 
         },
 
+        get_firstpage_url: function() {
+            // works:
+            // http://ops.epo.org/3.1/rest-services/published-data/images/EP/1000000/PA/firstpage.png?Range=1
+            // http://ops.epo.org/3.1/rest-services/published-data/images/US/20130311929/A1/thumbnail.tiff?Range=1
+            //var url_tpl = _.template('http://ops.epo.org/3.1/rest-services/published-data/images/<%= country %>/<%= docnumber %>/<%= kind %>/firstpage.png?Range=1');
+            var url_tpl = _.template('/api/ops/<%= country %><%= docnumber %>.<%= kind %>/image/firstdrawing');
+            var url = url_tpl({country: this['@country'], docnumber: this['@doc-number'], kind: this['@kind']});
+            return url;
+        },
+
 },
 
     select: function() {